% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_token.R
\name{add_token_to_renviron}
\alias{add_token_to_renviron}
\title{adds CHIMP_TOKEN to your environ file for secure communication with data chimp servers}
\usage{
add_token_to_renviron(token)
}
\description{
adds CHIMP_TOKEN to your environ file for secure communication with data chimp servers
}
